When copy used in place of rename (as a fall back workaround), it should be 'safe' copy, that attempts to rename a file within the same folder before doing non-atomic copy.

+++ svnkit/src/main/java/org/tmatesoft/svn/core/internal/wc/SVNFileUtil.java
@@ -515,7 +515,10 @@ public class SVNFileUtil {
         if (!isWindows) {
             renamed = src.renameTo(dst);
             if (!renamed && src.isFile() && !dst.exists()) {
-                copyFile(src, dst, false);
+                copyFile(src, dst, true);
+                if (!dst.isFile()) {
+                    copyFile(src, dst, false);
+                }
                 boolean deleted = deleteFile(src);
                 renamed = deleted && dst.isFile();
             }
